% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salesTS-package.R
\docType{package}
\name{salesTS-package}
\alias{salesTS}
\alias{salesTS-package}
\title{salesTS: Analiza Sprzedazy Detalicznej w Czasie (Retail Sales Time Series)}
\description{
Narzedzia do wczytywania, walidacji, czyszczenia, analizy metrycznej, wizualizacji oraz prognozowania detalicznych szeregow czasowych sprzedazy (na wzor zbioru Corporacion Favorita). Pakiet dostarcza spojny pipeline: od surowych plikow CSV, przez kontrole jakosci danych i inzynierie metryk, az po podsumowania menedzerskie i prognozy ARIMA oraz Prophet. Dokumentacja funkcji jest po angielsku, natomiast komunikaty dla uzytkownika i raporty biznesowe sa po polsku.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/example/salesTS}
  \item Report bugs at \url{https://github.com/example/salesTS/issues}
}

}
\author{
\strong{Maintainer}: Zespol Analityczny \email{analytics@example.com}

Authors:
\itemize{
  \item Zespol Analityczny \email{analytics@example.com}
}

}
\keyword{internal}
